+++ src/Compnents/Navbar.js
@@ -1,57 +1,16 @@
-import React, { useState } from 'react';
-import jaranda from 'Assets/jarandalogo.png';
-import {
-  NavbarContainer,
-  NavbarInnerContainer,
-  NavLink,
-  NavMenu,
-} from 'Styles/NavbarStyles';
+import React from 'react';
+import GuestNavbar from './GuestNavbar';
+import { NavbarContainer } from 'Styles/NavbarStyles';
+import UserNavbar from './UserNavbar';
 
 const Navbar = () => {
-  const [isActive, setIsActive] = useState('false');
-
-  const handleToggle = () => {
-    setIsActive(isActive);
-  };
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
 
   return (
     <div>
       <NavbarContainer>
-        <NavbarInnerContainer>
-          <NavLink to="/">
-            <img src={jaranda} alt="logo" />
-          </NavLink>
-          <NavMenu>
-            <NavLink
-              to="/parent"
-              onClick={handleToggle}
-              className={isActive ? null : 'active'}
-            >
-              자란다선생님 보기
-            </NavLink>
-            <NavLink
-              to="/teacher"
-              onClick={handleToggle}
-              className={isActive ? null : 'active'}
-            >
-              선생님지원하기
-            </NavLink>
-            <NavLink
-              to="/help"
-              onClick={handleToggle}
-              className={isActive ? null : 'active'}
-            >
-              이용안내
-            </NavLink>
-            <NavLink
-              to="/signin"
-              onClick={handleToggle}
-              className={isActive ? null : 'active'}
-            >
-              로그인/회원가입
-            </NavLink>
-          </NavMenu>
-        </NavbarInnerContainer>
+        {isLoggedIn ? <GuestNavbar /> : <UserNavbar />}
+        <GuestNavbar />
       </NavbarContainer>
     </div>
   );
